pyproject.toml
[project]
name = "object_extractor"
version = "0.1.0"
description = "Extract and process object outlines from images"
requires-python = ">=3.8"
dependencies = [
    "torch>=2.0.0",
    "numpy",
    "opencv-python",
    "segment-anything",
    "pillow",
    "typer",
    "supervision", 
    "groundingdino-py"  
]

[tool.setuptools.package-data]
object_extractor = ["configs/*"]